+++ api/routes/products.js
@@ -4,6 +4,7 @@ const Product = require("../models/product");
 const mongoose = require("mongoose");
 const multer = require("multer");
 const checkAuth = require('../middleware/check-auth');
+const ProductsController = require('../controllers/products');
 
 const storage = multer.diskStorage({
   destination: function (req, file, cb) {
@@ -32,134 +33,14 @@ const upload = multer({
   fileFilter: fileFilter,
 });
 
-router.get("/", checkAuth, (req, res, next) => {
-  Product.find()
-    .select("name price _id productImage")
-    .exec()
-    .then((docs) => {
-      const response = {
-        count: docs.length,
-        products: docs.map((doc) => {
-          return {
-            name: doc.name,
-            price: doc.price,
-            productImage: doc.productImage,
-            _id: doc._id,
-            request: {
-              type: "GET",
-              url: "http://localhost:3000/products" + doc._id,
-            },
-          };
-        }),
-      };
-      res.status(200).json(response);
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json({
-        error: err,
-      });
-    });
-});
+router.get("/", checkAuth, ProductsController.product_get_all);
 
-router.post("/", checkAuth, upload.single("productImage"), (req, res, next) => {
-  console.log(req.file);
-  const product = new Product({
-    _id: new mongoose.Types.ObjectId(),
-    name: req.body.name,
-    price: req.body.price,
-    productImage: req.file.path,
-  });
+router.post("/", checkAuth, upload.single("productImage"), ProductsController.product_create);
 
-  product
-    .save()
-    .then((result) => {
-      console.log(result);
-      res.status(200).json({
-        message: "Product Creadted Successfully",
-        createdProduct: {
-          name: result.name,
-          price: result.price,
-          _id: result._id,
-          request: {
-            type: "POST",
-            url: "http://localhost:3000/products/" + result._id,
-          },
-        },
-      });
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json({ error: err });
-    });
-});
+router.get("/:productId", checkAuth, ProductsController.product_delete_id);
 
-router.get("/:productId", checkAuth, (req, res, next) => {
-  const id = req.params.productId;
+router.patch("/:productId", checkAuth, ProductsController.product_update);
 
-  Product.findById(id)
-    .select("name price _id productImage")
-    .exec()
-    .then((doc) => {
-      console.log("From Database", doc);
-      if (doc) {
-        res.status(200).json({
-          product: doc,
-          request: {
-            type: "GET",
-            url: "http://localhost/products",
-          },
-        });
-      } else {
-        res.status(404).json({ message: "No valid Product Id found" });
-      }
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json({ error: err });
-    });
-});
-
-router.patch("/:productId", checkAuth, (req, res, next) => {
-  const id = req.params.productId;
-  const updateOps = {};
-  for (const ops of req.body) {
-    updateOps[ops.propName] = ops.value;
-  }
-  // Product.update({ _id: id }, {$set: {name: req.body.newName, price: req.body.newPrice}})
-  Product.update({ _id: id }, { $set: updateOps })
-    .exec()
-    .then((result) => {
-      console.log(result);
-      res.status(200).json({
-        messsage: "Product updated",
-        request: {
-          type: "GET",
-          url: "http://localhost:3000/products/" + id,
-        },
-      });
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json({
-        error: err,
-      });
-    });
-});
-
-router.delete("/:productId", checkAuth, (req, res, next) => {
-  const id = req.params.productId;
-  Product.remove({ _id: id })
-    .exec()
-    .then((result) => {
-      res.status(200).json(result);
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json({
-        error: err,
-      });
-    });
-});
+router.delete("/:productId", checkAuth, ProductsController.product_delete_id);
 
 module.exports = router;
