+++ api/routes/user.js
@@ -5,111 +5,12 @@ const User = require("../models/user");
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
 
-router.post("/signup", (req, res, next) => {
-  User.find({ email: req.body.email })
-    .exec()
-    .then((user) => {
-      if (user.length >= 1) {
-        return res.status(409).json({
-          message: "User with this email address has been created before.",
-        });
-      } else {
-        bcrypt.hash(req.body.password, 10, (err, hash) => {
-          if (err) {
-            return res.status(500).json({
-              error: err,
-            });
-          } else {
-            const user = new User({
-              _id: new mongoose.Types.ObjectId(),
-              email: req.body.email,
-              password: hash,
-              name: req.body.name,
-            });
-            user
-              .save()
-              .then((result) => {
-                console.log(result);
-                res.status(201).json({
-                  message: "User Created",
-                  data: {
-                    name: result.name,
-                    email: result.email,
-                    id: result._id,
-                  },
-                });
-              })
-              .catch((err) => {
-                console.log(result);
-                res.status(500).json({
-                  error: EvalError,
-                });
-              });
-          }
-        });
-      }
-    });
-});
+const UsersController = require('../controllers/users');
 
-router.post("/login", (req, res, next) => {
-  User.find({ email: req.body.email })
-    .exec()
-    .then((user) => {
-      if (user.length < 1) {
-        return res.status(401).json({
-          message: "Auth Failed",
-        });
-      }
-      bcrypt.compare(req.body.password, user[0].password, (err, result) => {
-        if (err) {
-          return res.status(401).json({
-            message: "Auth Failed",
-          });
-        }
-        if (result) {
-          const token = jwt.sign(
-            {
-              email: user[0].email,
-              userId: user[0]._id,
-            },
-            process.env.JWT_KEY,
-            {
-              expiresIn: "1h",
-            }
-          );
-          return res.status(200).json({
-            message: "Auth Successful",
-            token: token,
-          });
-        }
+router.post("/signup", UsersController.users_signup);
 
-        res.status(401).json({
-          message: "Auth Failed",
-        });
-      });
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json({
-        error: err,
-      });
-    });
-});
+router.post("/login", UsersController.users_login);
 
-router.delete("/:userId", (req, res, next) => {
-  User.remove({ _id: req.params.id })
-    .exec()
-    .then((result) => {
-      res.status(200).json({
-        message: "User Deleted",
-      });
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json({
-        error: err,
-      });
-    });
-});
+router.delete("/:userId", UsersController.users_delete);
 
 module.exports = router;
